重新调整组件的加载机制
