+++ src/cookie/Http.php
@@ -10,9 +10,10 @@
 namespace nb\cookie;
 
 use nb\Config;
+use nb\Pool;
 
 /**
- * Native
+ * Swoole
  *
  * @package nb\cookie
  * @link https://nb.cx
@@ -20,9 +21,13 @@ use nb\Config;
  * @author: collin <collin@nb.cx>
  * @date: 2017/11/28
  */
-class Native extends Driver {
+class Http extends Driver {
 
     protected $config;
+
+    protected $request;
+
+    protected $response;
 
     /**
      * Cookie初始化
@@ -37,6 +42,10 @@ class Native extends Driver {
         if (!empty($this->config['httponly'])) {
             ini_set('session.cookie_httponly', 1);
         }
+        $this->request  = Pool::value('\swoole\http\Request');
+        $this->response = Pool::value('\swoole\http\Response');
+        //$this->request  = \nb\driver\Swoole::$o->request;
+        //$this->response = \nb\driver\Swoole::$o->response;
     }
 
     /**
@@ -61,7 +70,6 @@ class Native extends Driver {
      * @return mixed
      */
     public function set($name, $value = '', $option = null) {
-
         // 参数设置(会覆盖黙认设置)
         if (!is_null($option)) {
             if (is_numeric($option)) {
@@ -81,11 +89,9 @@ class Native extends Driver {
             array_walk_recursive($value, 'self::jsonFormatProtect', 'encode');
             $value = 'nb:' . json_encode($value);
         }
-        $expire = !empty($config['expire']) ? $_SERVER['REQUEST_TIME'] + intval($config['expire']) : 0;
-        if ($config['setcookie']) {
-            setcookie($name, $value, $expire, $config['path'], $config['domain'], $config['secure'], $config['httponly']);
-        }
-        $_COOKIE[$name] = $value;
+        $expire = !empty($config['expire']) ? Request::ins()->requestTime + intval($config['expire']) : 0;
+        $this->response->cookie($name, $value, $expire, $config['path'], $config['domain'], $config['secure'], $config['httponly']);
+        $this->request->cookie[$name] = $value;
     }
 
     /**
@@ -112,7 +118,7 @@ class Native extends Driver {
     public function has($name, $prefix = null) {
         $prefix = !is_null($prefix) ? $prefix : $this->config['prefix'];
         $name = $prefix . $name;
-        return isset($_COOKIE[$name]);
+        return isset($this->response->cookie[$name]);
     }
 
     /**
@@ -124,8 +130,8 @@ class Native extends Driver {
     public function get($name, $prefix = null) {
         $prefix = !is_null($prefix) ? $prefix : $this->config['prefix'];
         $name = $prefix . $name;
-        if (isset($_COOKIE[$name])) {
-            $value = $_COOKIE[$name];
+        if (isset($this->request->cookie[$name])) {
+            $value = $this->request->cookie[$name];
             if (0 === strpos($value, 'nb:')) {
                 $value = substr($value, 6);
                 $value = json_decode($value, true);
@@ -147,10 +153,18 @@ class Native extends Driver {
         $prefix = is_null($prefix) ? $config['prefix'] : $prefix;
         $name = $prefix . $name;
         if ($config['setcookie']) {
-            setcookie($name, '', $_SERVER['REQUEST_TIME'] - 3600, $config['path'], $config['domain'], $config['secure'], $config['httponly']);
+            $this->response->cookie(
+                $name,
+                '',
+                Http::obj()->requestTime - 3600,
+                $config['path'],
+                $config['domain'],
+                $config['secure'],
+                $config['httponly']
+            );
         }
         // 删除指定cookie
-        unset($_COOKIE[$name]);
+        unset($this->response->cookie[$name]);
     }
 
     /**
@@ -160,7 +174,7 @@ class Native extends Driver {
      */
     public function clear($prefix = '') {
         // 清除指定前缀的所有cookie
-        if (empty($_COOKIE)) {
+        if (empty($this->request->cookie)) {
             return;
         }
         // 要删除的cookie前缀，不指定则删除config设置的指定前缀
@@ -170,15 +184,18 @@ class Native extends Driver {
         foreach ($_COOKIE as $key => $val) {
             $key = $prefix.$key;
             if ($config['setcookie']) {
-                setcookie($key, '', $_SERVER['REQUEST_TIME'] - 3600, $config['path'], $config['domain'], $config['secure'], $config['httponly']);
+                $this->response->cookie(
+                    $key,
+                    '',
+                    Http::obj()->requestTime - 3600,
+                    $config['path'],
+                    $config['domain'],
+                    $config['secure'],
+                    $config['httponly']
+                );
             }
-            unset($_COOKIE[$key]);
+            unset($this->request->cookie[$key]);
         }
     }
 
-    private function jsonFormatProtect(&$val, $key, $type = 'encode') {
-        if (!empty($val) && true !== $val) {
-            $val = 'decode' == $type ? urldecode($val) : urlencode($val);
-        }
-    }
 }
