+++ src/debug/Base.php
@@ -1,0 +1,23 @@
+<?php
+/*
+ * This file is part of the NB Framework package.
+ *
+ * Copyright (c) 2018 https://nb.cx All rights reserved.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace nb\debug;
+
+/**
+ * Base
+ *
+ * @package nb\src\debug
+ * @link https://nb.cx
+ * @author: collin <collin@nb.cx>
+ * @date: 2018/8/7
+ */
+class Base extends Command {
+
+
+}
