+++ src/dispatcher/Http.php
@@ -14,6 +14,7 @@ use nb\Request;
 use nb\Response;
 use nb\Router;
 use nb\Pool;
+use nb\src\dispatcher\Php;
 
 /**
  * Swoole
@@ -24,7 +25,7 @@ use nb\Pool;
  * @author: collin <collin@nb.cx>
  * @date: 2017/12/3
  */
-class Swoole extends Native {
+class Http extends Php {
 
     protected $config = [
         'path'=>__APP__,
