+++ src/request/Base.php
@@ -1,0 +1,23 @@
+<?php
+/*
+ * This file is part of the NB Framework package.
+ *
+ * Copyright (c) 2018 https://nb.cx All rights reserved.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace nb\request;
+
+/**
+ * Native
+ *
+ * @package nb\request
+ * @link https://nb.cx
+ * @author: collin <collin@nb.cx>
+ * @date: 2017/11/28
+ */
+class Base extends Php {
+
+
+}
